pyproject.toml
[tool.poetry]
name = "praetorian-test-deploy"
version = "0.1.0"
description = ""
authors = ["Adam Žúrek <adamzurek14@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.15.0"
fabric = "^2.5.0"
praetorian-fabric = "^0.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
